split `Order.ts` & renames & import fixes
